# Speedup `to_hex` (~2x faster)