## Which issue does this PR close?

N/A

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

We can speedup `to_hex` by writing string values directly to the string array, instead of making temporary allocations.

```
to_hex i32 array: 1024  time:   [14.239 µs 14.295 µs 14.354 µs]
                        change: [-48.522% -48.285% -48.031%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

to_hex i64 array: 1024  time:   [15.471 µs 15.546 µs 15.624 µs]
                        change: [-46.390% -46.208% -46.007%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

- Add a benchmark
- Avoid string allocations

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

Relying on existing tests.

## Are there any user-facing changes?

Faster execution.

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->